Function mocking, multiline support and bug fixes
The order of the decorations could not match the order of execution. This is purely due to how vs code handles decorations (i.e. not in the order they are in the array you give) so I had no choice but to fuse them all into a single one.
Also fixed a bug where errors within a trace were displayed at the first mentioned line, which is less accurate